backend/requirements.txt
flask
requests
flask_cors
argon2-cffi